+++ xill-maven-plugin/src/main/java/nl/xillio/xill/maven/services/FileSystemFactory.java
@@ -15,18 +15,17 @@
  */
 package nl.xillio.xill.maven.services;
 
-import org.codehaus.plexus.archiver.FileSet;
-import org.codehaus.plexus.archiver.util.DefaultFileSet;
+import java.io.IOException;
+import java.net.URI;
+import java.nio.file.FileSystem;
+import java.nio.file.FileSystems;
+import java.util.Map;
 
-import java.io.File;
-import java.nio.file.Path;
-
-public class FileSetFactory {
-    public static final String ROBOTS_DIRECTORY = "robots" + File.separator;
-
-    public FileSet createFileSet(Path directory) {
-        DefaultFileSet fileSet = new DefaultFileSet(directory.toFile());
-        fileSet.setPrefix(ROBOTS_DIRECTORY);
-        return fileSet;
+/**
+ * Created by Dwight.Peters on 14-Jul-17.
+ */
+public class FileSystemFactory {
+    public FileSystem createFileSystem(URI uri, Map<String, String> env) throws IOException {
+        return FileSystems.newFileSystem(uri, env);
     }
 }