+++ xill-processor/src/main/java/nl/xillio/xill/components/operators/FromList.java
@@ -19,6 +19,7 @@ import nl.xillio.xill.api.Debugger;
 import nl.xillio.xill.api.components.*;
 import nl.xillio.xill.api.errors.NotImplementedException;
 import nl.xillio.xill.api.errors.RobotRuntimeException;
+import nl.xillio.xill.components.expressions.VariableAccessExpression;
 
 import java.util.Arrays;
 import java.util.Collection;
@@ -39,14 +40,14 @@ public class FromList implements Processable {
     }
 
     @Override
-    public InstructionFlow<MetaExpression> process(final Debugger debugger) throws RobotRuntimeException {
+    public InstructionFlow<MetaExpression> process(final Debugger debugger) {
 
         MetaExpression listMeta = this.list.process(debugger).get();
         MetaExpression indexMeta = this.index.process(debugger).get();
         listMeta.registerReference();
         indexMeta.registerReference();
 
-        InstructionFlow<MetaExpression> result = process(listMeta, indexMeta, debugger);
+        InstructionFlow<MetaExpression> result = process(listMeta, indexMeta);
 
         result.get().preventDisposal();
 
@@ -58,10 +59,7 @@ public class FromList implements Processable {
         return result;
     }
 
-    @SuppressWarnings({
-            "unchecked"
-    })
-    private InstructionFlow<MetaExpression> process(MetaExpression listMeta, MetaExpression indexMeta, Debugger debugger) {
+    private InstructionFlow<MetaExpression> process(MetaExpression listMeta, MetaExpression indexMeta) {
 
         if (indexMeta.getType() != ExpressionDataType.ATOMIC || indexMeta.isNull()) {
             return InstructionFlow.doResume(ExpressionBuilderHelper.NULL);
@@ -69,27 +67,52 @@ public class FromList implements Processable {
 
         switch (listMeta.getType()) {
             case LIST:
-                if (Double.isNaN(indexMeta.getNumberValue().doubleValue())) {
-                    throw new RobotRuntimeException("The list does not contain any element called '" + indexMeta.getStringValue() + "' (a list does not have named elements).");
-                }
-                List<MetaExpression> list = listMeta.getValue();
-                int index = indexMeta.getNumberValue().intValue();
-                if (index < 0 || index >= list.size()) {
-                    throw new RobotRuntimeException("Illegal value for list index: " + index);
-                }
-                MetaExpression listResult = list.get(index);
-                return InstructionFlow.doResume(listResult);
+                return getMetaExpressionListInstructionFlow(listMeta, indexMeta);
             case OBJECT:
-                MetaExpression objectResult = ((Map<String, MetaExpression>) listMeta.getValue()).get(indexMeta.getStringValue());
-                if (objectResult == null) {
-                    return InstructionFlow.doResume(ExpressionBuilderHelper.NULL);
-                }
-                return InstructionFlow.doResume(objectResult);
+                return getMetaExpressionObjectInstructionFlow(listMeta, indexMeta);
             case ATOMIC:
-                throw new RobotRuntimeException("Cannot get member of ATOMIC value.");
+                String message = getMemberFromAtomicErrorMessage(listMeta, indexMeta);
+                throw new RobotRuntimeException(message);
             default:
                 throw new NotImplementedException("This type has not been implemented.");
         }
+    }
+
+    private String getMemberFromAtomicErrorMessage(MetaExpression listMeta, MetaExpression indexMeta) {
+        String message = "Cannot get member '" + indexMeta.getStringValue() + "' from ATOMIC value";
+
+        // Check if the list is a variable, add the name to the message.
+        if (list instanceof VariableAccessExpression) {
+            message += " '" + ((VariableAccessExpression)list).getVariableName() + "'";
+        }
+
+        // Check if the list is null.
+        if (listMeta.isNull()) {
+            message += " (value is null)";
+        }
+
+        return message + ".";
+    }
+
+    private InstructionFlow<MetaExpression> getMetaExpressionObjectInstructionFlow(MetaExpression listMeta, MetaExpression indexMeta) {
+        MetaExpression objectResult = listMeta.<Map<String, MetaExpression>>getValue().get(indexMeta.getStringValue());
+        if (objectResult == null) {
+            return InstructionFlow.doResume(ExpressionBuilderHelper.NULL);
+        }
+        return InstructionFlow.doResume(objectResult);
+    }
+
+    private InstructionFlow<MetaExpression> getMetaExpressionListInstructionFlow(MetaExpression listMeta, MetaExpression indexMeta) {
+        if (Double.isNaN(indexMeta.getNumberValue().doubleValue())) {
+            throw new RobotRuntimeException("The list does not contain any element called '" + indexMeta.getStringValue() + "' (a list does not have named elements).");
+        }
+        List<MetaExpression> listMetas = listMeta.getValue();
+        int intIndex = indexMeta.getNumberValue().intValue();
+        if (intIndex < 0 || intIndex >= listMetas.size()) {
+            throw new RobotRuntimeException("Illegal value for list index: " + intIndex);
+        }
+        MetaExpression listResult = listMetas.get(intIndex);
+        return InstructionFlow.doResume(listResult);
     }
 
     @Override