+++ xill-processor/src/main/java/nl/xillio/xill/plugins/date/constructs/IsDateConstruct.java
@@ -30,11 +30,13 @@ public class IsDateConstruct extends BaseDateConstruct {
 
     @Override
     public ConstructProcessor prepareProcess(final ConstructContext context) {
-        return new ConstructProcessor(IsDateConstruct::process,
-                new Argument("date", ATOMIC));
+        return new ConstructProcessor(
+                this::process,
+                new Argument("date", ATOMIC)
+        );
     }
 
-    static MetaExpression process(final MetaExpression dateVar) {
+    private MetaExpression process(final MetaExpression dateVar) {
         return fromValue(dateVar.getMeta(Date.class) != null);
     }
 }