+++ xill-processor/src/test/java/nl/xillio/xill/plugins/date/constructs/InfoConstructTest.java
@@ -15,6 +15,7 @@
  */
 package nl.xillio.xill.plugins.date.constructs;
 
+import nl.xillio.xill.TestUtils;
 import nl.xillio.xill.api.components.MetaExpression;
 import nl.xillio.xill.plugins.date.services.DateService;
 import org.testng.annotations.Test;
@@ -24,7 +25,7 @@ import java.time.ZonedDateTime;
 import java.util.HashMap;
 import java.util.Map;
 
-import static nl.xillio.xill.plugins.date.utils.MockUtils.mockDateExpression;
+import static nl.xillio.xill.plugins.date.constructs.IsBeforeConstructTest.createDateTimeExpression;
 import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.*;
 import static org.testng.Assert.assertEquals;
@@ -34,33 +35,35 @@ import static org.testng.Assert.assertEquals;
  *
  * @author Geert Konijnendijk
  */
-public class InfoConstructTest {
+public class InfoConstructTest extends TestUtils {
 
     /**
      * Test the process method and see if it returns the correct keys in the returned map
      */
     @Test
     public void testProcess() {
+        InfoConstruct infoConstruct = new InfoConstruct();
+
         // Mock
         DateService dateService = mock(DateService.class);
         Map<String, Long> fieldValues = new HashMap<>();
-        fieldValues.put("Field1", 10l);
-        fieldValues.put("Field2", 20l);
-        fieldValues.put("Field3", 30l);
+        fieldValues.put("Field1", 10L);
+        fieldValues.put("Field2", 20L);
+        fieldValues.put("Field3", 30L);
         when(dateService.getFieldValues(any())).thenReturn(fieldValues);
         ZoneId zoneId = mock(ZoneId.class);
         when(zoneId.toString()).thenReturn("Europe/Amsterdam");
         when(dateService.getTimezone(any())).thenReturn(zoneId);
-        boolean isInFuture = true, isInPast = false;
-        ;
-        when(dateService.isInFuture(any())).thenReturn(isInFuture);
-        when(dateService.isInPast(any())).thenReturn(isInPast);
+
+        when(dateService.isInFuture(any())).thenReturn(true);
+        when(dateService.isInPast(any())).thenReturn(false);
         // ZonedDateTime is final, don't mock
         ZonedDateTime date = ZonedDateTime.now();
-        MetaExpression dateExpression = mockDateExpression(date);
+        MetaExpression dateExpression = createDateTimeExpression(date);
+        infoConstruct.setDateService(dateService);
 
         // Run
-        MetaExpression info = InfoConstruct.process(dateExpression, dateService);
+        MetaExpression info = process(infoConstruct, dateExpression);
 
         // Verify
         verify(dateService).getFieldValues(any());
@@ -69,10 +72,10 @@ public class InfoConstructTest {
         verify(dateService).isInPast(any());
 
         // Assert
-        Map<String, MetaExpression> infoMap = (Map<String, MetaExpression>) info.getValue();
+        Map<String, MetaExpression> infoMap = info.getValue();
         fieldValues.forEach((k, v) -> assertEquals(infoMap.get(k).getNumberValue().longValue(), (long) v));
         assertEquals(infoMap.get("timeZone").getStringValue(), zoneId.toString());
-        assertEquals(infoMap.get("isInFuture").getBooleanValue(), isInFuture);
-        assertEquals(infoMap.get("isInPast").getBooleanValue(), isInPast);
+        assertEquals(infoMap.get("isInFuture").getBooleanValue(), true);
+        assertEquals(infoMap.get("isInPast").getBooleanValue(), false);
     }
 }